\name{otodom_ab-package}
\alias{otodom_ab-package}
\alias{otodom_ab}
\docType{package}
\title{
\packageTitle{otodom_ab}
}
\description{
\packageDescription{otodom_ab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{otodom_ab}
\packageIndices{otodom_ab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{otodom_ab}

Maintainer: \packageMaintainer{otodom_ab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
